Ok update again API UpdateStudent

+++ src/main/java/jpa/systemstudent/dto/StudentDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor
 public class StudentDto {
-    private int studentId;
+    private Long studentId;
 
     private String studentName;
